% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_df}
\alias{covid_df}
\title{Real COVID-19 Data from Our World in Data}
\format{
\code{"covid_df"} is a data frame with 306,323 cases (rows) and 67 variables
(columns).
}
\source{
\href{https://ourworldindata.org/coronavirus}{Our World in Data}
}
\usage{
data(covid_df)
}
\description{
This data set gives real COVID data of US and UK. The COVID related measurements include
daily new cases, total confirmed cases, reproduction_rate etc and the measuring period covers
from 2020-01-03 to 2023-04-26. This data set also includes country specific information, like population.
}
\examples{
data(covid_df)
}
\references{
Edouard Mathieu, Hannah Ritchie, Lucas Rodés-Guirao, Cameron Appel, Charlie Giattino,
Joe Hasell, Bobbie Macdonald, Saloni Dattani, Diana Beltekian, Esteban Ortiz-Ospina and Max Roser (2020) - "Coronavirus Pandemic (COVID-19)".
}
\keyword{datasets}
