% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_tuning.R
\name{parameter_tuning}
\alias{parameter_tuning}
\title{Tuning covid spread model using real world data sets}
\usage{
parameter_tuning(range1, range2, df, parallel = TRUE, ...)
}
\arguments{
\item{range1}{muT tuning ranges}

\item{range2}{sizeV tuning ranges}

\item{df}{real world data set of a specific country}

\item{parallel}{a Boolean variable, indicating whether a parallel processing is wanted, the default value is TRUE.}

\item{...}{Other arguments passed on to methods.}
}
\value{
the function returns a data frame of the best value of muT, sizeV, and the minimized difference.
}
\description{
This function uses real world data set to tune covid spread model, aiming to minimize the difference between simulated total number of cases and actutal total number of cases.
}
