% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_func.R
\name{TransSimu}
\alias{TransSimu}
\title{Simulate the spread of COVID disease among people}
\usage{
TransSimu(
  days = 300,
  nd = 30,
  Rt = rr,
  muT = 4,
  sizeV = 1,
  limit = 1000000,
  pp = 0.001,
  n0 = 1
)
}
\arguments{
\item{days}{observation period}

\item{nd}{simulation period}

\item{Rt}{infection rate pattern}

\item{muT}{the mean time an infected person will transmit the virus to another person}

\item{sizeV}{the dispersion parameter so that variance = mu + mu^2/size}

\item{limit}{the target/study population size}

\item{pp}{the proportion of people with immunity in the population}

\item{n0}{the initial number of infectious persons}
}
\value{
the function returns a list of daily active cases number, daily new cases number and daily total number of cases over the observation period.
}
\description{
This function gives out the distributions of certain measurements that help evaluate COVID situation. These measurements include daily new cases, active cases etc.
}
\examples{
rr = c(rep(3.4,10),rep(3.1,10),rep(2.2,5),rep(1.7,4),rep(1.4,6),rep(1.2,6),rep(1.1,4),rep(1,8),rep(0.9,7),rep(0.6,10),rep(0.1,30))
TransSimu(days = 100, nd = 30, Rt = rr)

}
