% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_covid.R
\name{bootstrap_run}
\alias{bootstrap_run}
\title{Bootstrap in covid simulation}
\usage{
bootstrap_run(n, country, covid_df)
}
\arguments{
\item{n}{number of times to bootstrap}

\item{country}{country of interest}

\item{covid_df}{the real world covid data set}
}
\value{
the function returns a list of daily active cases number, daily new cases number after bootstrapping.
}
\description{
This function helps to do bootstrap. 'dplyr' package is used for bootstrap data preparation.
}
